third_party/blink/web_tests/http/tests/security/isolatedWorld/bypass-main-world-csp-iframes-expected.txt

ALERT: Running test #1

CONSOLE ERROR: Framing 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=1' violates the following Content Security Policy directive: "frame-src 'none'". The request has been blocked.

ALERT: Running test #2

CONSOLE ERROR: Framing 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=2' violates the following Content Security Policy directive: "frame-src 'none'". The request has been blocked.

ALERT: Running test #3

ALERT: Starting to bypass main world's CSP:
ALERT: iframe loaded in test #3
ALERT: Running test #4

CONSOLE ERROR: Framing 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=4' violates the following Content Security Policy directive: "frame-src 'none'". The request has been blocked.

ALERT: iframe loaded in test #4.5
ALERT: Running test #5

CONSOLE ERROR: Framing 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=5.5' violates the following Content Security Policy directive: "frame-src 'none'". The request has been blocked.

ALERT: Running test #6

CONSOLE ERROR: Framing 'http://127.0.0.1:8000/security/isolatedWorld/resources/iframe.php?test=6' violates the following Content Security Policy directive: "frame-src 'none'". The request has been blocked.

ALERT: Running test #7

 This test ensures that iframes loaded from isolated worlds marked with their own Content Security Policy aren't affected by the page's content security policy. Extensions, for example, should be able to load any resource they like.
